ImportantTaskTests.swift, RegularTaskTests.swift
Разделил проверку задач на 2 файла, отказался от отдельного енама со стабом. Отказался от rawValue, вместо него используем dayCountForCompletion. Изменил логику проверки regularTask
